requirements.txt
asgiref==3.9.2
Django==4.2
mysql-connector-python==9.4.0
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
